wip, we now have monomials
